simplify delete confirmation
