--- src/apps/content-editor/src/app/views/LinkEdit/LinkDeleteConfirmation.less
@@ -1,4 +1,0 @@
-.ModalFooter {
-  display: flex;
-  justify-content: space-between;
-}
