+++ src/apps/content-editor/src/app/views/LinkEdit/LinkEdit.less
@@ -1,5 +1,6 @@
 .Editor {
   width: 85vw;
+  height: calc(100vh - 54px);
   .LinkEdit {
     width: 800px;
     .Row {
